Adding Automatic Speaker Recognition to Summer-Intern-Projects
Desc: The project Automatic Speaker Recognition (referred to as ASR) is
designed to recognize and distinguish between multiple voices of
different people. A large dataset of various speakers is used and each
audio sample is tested with the existing audio samples and matched with
the one whose features are the closest. This is achieved with the help
of MFCC, LPC, and LBG algorithms.

Team Members: @Jagathveerendra, @DweejaReddy, Ayush Varma, Gautam
Tahilyani, Aditya Undrikar
Mentors: @sibam23, @ThanmayJ, @LuqmanFarooqui